feat(tests): cut the matrix over to the generated backend (Phase 5)
matrix.ts is now the single Combo[] source: the backend core (server / data /
db / orm / auth on the home deploy) comes from generateMatrix(), and a `residue`
of suites holds everything the generator doesn't produce — the peripheral
feature suites (UI / CSS / analytics / sentry / storybook / prettier / aws /
skills / linter-strip) and the deploy-specific backend variants (Cloudflare D1,
dokploy, auth0 / better-auth cloud) extracted from the old data/auth suites.

buildCombos is gone — its flatten+dedup folded into matrix.ts (which also breaks
the combos↔matrix import cycle); runner.ts consumes the matrix directly. The
migration-only matrix-diff spec becomes matrix.local.spec.ts: no duplicate
combos, and the backend core covers every satisfiable interaction pair (t=2).

73 combos (was 106): the 40 pairwise-redundant home backend combos drop for the
same t=2 coverage; deploy-specifics are preserved. A full test:e2e run is the
remaining validation before merge.
